mysite/requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.19
django-phone-field==1.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
Faker==7.0.1
idna==2.10
importlib-metadata==3.10.0
Markdown==3.3.4
Pillow==8.1.2
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
qiniu==7.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
